praxis.core - FIX PMap getDouble default value should be double not int.

+++ praxis.core/src/net/neilcsmith/praxis/core/types/PMap.java
@@ -82,7 +82,7 @@ public class PMap extends Argument {
         return def;
     }
 
-    public double getDouble(String key, int def) {
+    public double getDouble(String key, double def) {
         Argument val = get(key);
         if (val != null) {
             try {
